requirement.txt
boto==2.43.0
bz2file==0.98
chainer==1.18.0
Cython==0.25.1
filelock==2.0.7
gensim==0.13.3
mecab-python3==0.7
nose==1.3.7
numpy==1.11.2
protobuf==3.1.0.post1
requests==2.12.0
scikit-learn==0.18
scipy==0.18.1
six==1.10.0
smart-open==1.3.5
